Getting Process information
Implemented getting PIDs of running processes.
Added each process to a vector in System.
For each process, Uid and User are obtained and displayed.

Additional:
Think about creating each process object on the heap.
Store a vector of pointers to each object.
This will prevent a potential stack overflow at the cost of performance.

Possible bugs:
The cursor flashes back and forth when running the program.