Process object refactor
Now when a process object is created, all of the process's functions are
called and all information is tored in it's object member variables.
This makes sorting faster and prevents segmentation faults that occur
when trying to get information about a pid/process that is no longer
running.
Sorting is now done w.r.t RAM instead of CPU utilization.

In Progress:
Process RAM and CPU values are stil being occasionally displayed wrong.
This may be a issue with ncurses letting old data persist on the screen.
CPU sometimes shows up as "-nan" or infinity.
RAM still shows up as ###.###.## occasionally.